requirements/backend.txt
# pip install -r requirements/backend.txt


fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow
jinja2
